Added the vertical step! 👨‍🔬
I've written the SetIndivBends function, which tunes the notes in a chord individually to just intonation! That function needs text files that define the correct interval sizes, so I've written some of those as well in TuningTables.

Chords, like Notes, are now defined with a note index from 0-11 instead of a string name.